Added tranfer of money and changed model of storing accounts
Now we store in the db the list of all the transfers of
money that have been done for an account instead of the current
amount.

Is not to performant in terms of storage but we don't care and
it allows to keep the history of transfers for a user.